Independent JsonSerializers for the Input and Output formatters and a MvcBuilder extension to simplify configuration.
Because JSON.NET will see the ResourceConverter.WriteJson() call to JsonSerializer.Serialize() as a reference loop (already marked as serialized),
must set the ReferenceLoopHandling to Serialize not Error. Both the input and output serializers are created from the same settings. This change
allows each formatter to create there own serializer and set the required settings accordingly.

Idealy this should be fixed by not using a JsonConverter on the Representation to setup the ConverterContext by move these up into the formatters.